requirements.txt
Cython
fastapi
h5py
numpy
opencv-python
pillow
pyserial
ruckig
scipy
setuptools
wheel
